./configure optionally reads the -tkdefs and -tklibs from the environment (if set)
This is useful for compiling against a non-default Tk when the flags -tkdefs/-tklibs cannot
be set explicitly for some reason. I.e. when installing via opam (and in particular on macOS
with a homebrew installation of tcl-tk).

For instance, do:
$ export TK_DEFS='-I/usr/local/Cellar/tcl-tk/8.6.13_5/include/tcl-tk'
$ export TK_LIBS='-L/usr/local/Cellar/tcl-tk/8.6.13_5/lib -ltk8.6 -ltkstub8.6 -ltcl8.6 -ltclstub8.6'

before
$ opam install labltk

See also https://github.com/garrigue/labltk/issues/18