Tcl 9 support (#28)
* configure: Detect Tcl & Tk 9.0
* build: Ignore temporary files more thoroughly
* support: Use const for arg list in Tcl_CmdProc
Tcl_CmdProc has used const by default for this arg list since at least
Tcl 8.4 (first released in 2002), so just use 'const'.  It is no
longer optional in Tcl 9.
* support: Use Tcl_* versions of events
Tk 8.6 (released in 2012) defines TK_*_EVENTS as aliases for
TCL_*_EVENTS, and Tcl 9 removes these aliases.  Similarly for
functions like Tcl_DoOneEvents etc.  Replace the aliases so this will
work on Tcl 9.
Note that TK_X_EVENTS is no longer a thing (even in Tcl 8.6).  In Tcl
8.6 it was an alias for TCL_WINDOW_EVENTS and in Tcl 9 it was removed
entirely.  I chose to remove it completely here.
* support: Remove obsolete _ANSI_ARGS_
* support: Use Tcl_Size in Tcl 9
In Tcl 9, a new typedef Tcl_Size must be used here.  It's important to
get the correct size since we pass a pointer to this type.  Use
TCL_SIZE_MAX as a sentinel since this macro is newly defined in Tcl 9.

Signed-off-by: Richard W.M. Jones <rjones@redhat.com>